PID: LuxFloat rewritten from scratch. New I-term back-tracking for anti-windup prevention
AIRMODE: Air mode anti-windup improvements

PID: Remove integer PID (leave only one PID controller), remove pid_controller setting, remove float-point PID parameters (scale integer parameters to use in float-point PIDC)

PID: New set of PID tuning multipliers for better tuning range

PID: Make max_angle_inclination a PID-profile variable

PID: Reinstate yaw_p_limit

PID: Switch to a noise-robust differentiator

Remove GTUNE

PID: Use same self-leveling strength for ANGLE and HORIZON

PID: Self-leveling LPF to smooth out leveling response to rapid attitude or target angle change

+++ src/main/common/filter.c
@@ -85,7 +85,7 @@ float filterApplyBiQuad(float sample, biquad_t *state)
 }
 
 // PT1 Low Pass filter (when no dT specified it will be calculated from the cycleTime)
-float filterApplyPt1(float input, filterStatePt1_t *filter, uint8_t f_cut, float dT)
+float filterApplyPt1(float input, filterStatePt1_t *filter, float f_cut, float dT)
 {
 	// Pre calculate and store RC
 	if (!filter->RC) {
